requirements.txt
flask>=3.0
requests>=2.31
psycopg[binary,pool]>=3.1
argon2-cffi>=23.1
gunicorn>=22.0
pydantic>=2.0
pydantic[email]>=2.0
postkit @ git+https://github.com/varunchopra/postkit.git#subdirectory=sdk
